refactor: make the requested changes
